Refactor code in type builders
