+++ compiler/ballerina-lang/src/main/java/io/ballerina/compiler/api/Types.java
@@ -20,7 +20,6 @@ package io.ballerina.compiler.api;
 import io.ballerina.compiler.api.impl.BallerinaTypes;
 import io.ballerina.compiler.api.impl.SymbolFactory;
 import io.ballerina.compiler.api.impl.symbols.TypesFactory;
-import io.ballerina.compiler.api.impl.types.TypeBuilder;
 import io.ballerina.compiler.api.symbols.Symbol;
 import io.ballerina.compiler.api.symbols.TypeSymbol;
 import org.wso2.ballerinalang.compiler.PackageCache;
