+++ compiler/ballerina-lang/src/main/java/io/ballerina/compiler/api/impl/BallerinaTypes.java
@@ -18,8 +18,7 @@
 package io.ballerina.compiler.api.impl;
 
 import io.ballerina.compiler.api.Types;
-import io.ballerina.compiler.api.impl.types.BallerinaTypeBuilder;
-import io.ballerina.compiler.api.impl.types.TypeBuilder;
+import io.ballerina.compiler.api.TypeBuilder;
 import io.ballerina.compiler.api.impl.util.FieldMap;
 import io.ballerina.compiler.api.symbols.Symbol;
 import org.ballerinalang.model.elements.PackageID;
