+++ compiler/ballerina-lang/src/main/java/io/ballerina/compiler/api/impl/type/builders/BallerinaMapTypeBuilder.java
@@ -16,17 +16,18 @@
  * under the License.
  */
 
-package io.ballerina.compiler.api.impl.types;
+package io.ballerina.compiler.api.impl.type.builders;
 
+import io.ballerina.compiler.api.TypeBuilder;
 import io.ballerina.compiler.api.impl.symbols.AbstractTypeSymbol;
 import io.ballerina.compiler.api.impl.symbols.TypesFactory;
-import io.ballerina.compiler.api.symbols.StreamTypeSymbol;
+import io.ballerina.compiler.api.symbols.MapTypeSymbol;
 import io.ballerina.compiler.api.symbols.TypeSymbol;
 import org.wso2.ballerinalang.compiler.semantics.model.SymbolTable;
 import org.wso2.ballerinalang.compiler.semantics.model.symbols.BTypeSymbol;
 import org.wso2.ballerinalang.compiler.semantics.model.symbols.SymTag;
 import org.wso2.ballerinalang.compiler.semantics.model.symbols.Symbols;
-import org.wso2.ballerinalang.compiler.semantics.model.types.BStreamType;
+import org.wso2.ballerinalang.compiler.semantics.model.types.BMapType;
 import org.wso2.ballerinalang.compiler.semantics.model.types.BType;
 import org.wso2.ballerinalang.compiler.util.CompilerContext;
 import org.wso2.ballerinalang.compiler.util.Names;
@@ -34,69 +35,52 @@ import org.wso2.ballerinalang.compiler.util.TypeTags;
 import org.wso2.ballerinalang.util.Flags;
 
 /**
- * The implementation of the methods used to build the Stream type descriptor in Types API.
+ * The implementation of the methods used to build the Map type descriptor in Types API.
  *
  * @since 2201.2.0
  */
-public class BallerinaStreamTypeBuilder implements TypeBuilder.STREAM {
+public class BallerinaMapTypeBuilder implements TypeBuilder.MAP {
 
     private final TypesFactory typesFactory;
     private final SymbolTable symTable;
-    private TypeSymbol valueType;
-    private TypeSymbol completionType;
+    private TypeSymbol typeParam;
 
-    public BallerinaStreamTypeBuilder(CompilerContext context) {
+    protected BallerinaMapTypeBuilder(CompilerContext context) {
         typesFactory = TypesFactory.getInstance(context);
         symTable = SymbolTable.getInstance(context);
     }
 
     @Override
-    public TypeBuilder.STREAM withValueType(TypeSymbol valueType) {
-        this.valueType = valueType;
+    public TypeBuilder.MAP withTypeParam(TypeSymbol typeParam) {
+        this.typeParam = typeParam;
         return this;
     }
 
     @Override
-    public TypeBuilder.STREAM withCompletionType(TypeSymbol completionType) {
-        this.completionType = completionType;
-        return this;
-    }
+    public MapTypeSymbol build() {
 
-    @Override
-    public StreamTypeSymbol build() {
-        BTypeSymbol streamSymbol = Symbols.createTypeSymbol(SymTag.TYPE, Flags.PUBLIC, Names.EMPTY,
+        BTypeSymbol mapTSymbol = Symbols.createTypeSymbol(SymTag.TYPE, Flags.PUBLIC, Names.EMPTY,
                 symTable.rootPkgSymbol.pkgID, null, symTable.rootPkgSymbol, symTable.builtinPos,
                 symTable.rootPkgSymbol.origin);
 
-        BStreamType streamType = new BStreamType(TypeTags.STREAM, getValueBType(this.valueType),
-                getCompletionBType(this.completionType), streamSymbol);
+        BMapType mapType = new BMapType(TypeTags.MAP, getBType(typeParam), mapTSymbol);
+        mapTSymbol.type = mapType;
+        MapTypeSymbol mapTypeSymbol = (MapTypeSymbol) typesFactory.getTypeDescriptor(mapType);
+        this.typeParam = null;
 
-        streamSymbol.type = streamType;
-
-        StreamTypeSymbol streamTypeSymbol = (StreamTypeSymbol) typesFactory.getTypeDescriptor(streamType);
-        this.valueType = null;
-        this.completionType = null;
-
-        return streamTypeSymbol;
+        return mapTypeSymbol;
     }
 
-    private BType getValueBType(TypeSymbol valueType) {
-        if (valueType == null) {
-            return symTable.anyType;
+    private BType getBType(TypeSymbol typeSymbol) {
+        if (typeSymbol == null) {
+            return null;
         }
 
-        if (valueType instanceof AbstractTypeSymbol) {
-            return ((AbstractTypeSymbol) valueType).getBType();
+        if (typeSymbol instanceof AbstractTypeSymbol
+                && typeSymbol.subtypeOf(typesFactory.getTypeDescriptor(symTable.anyType))) {
+            return ((AbstractTypeSymbol) typeSymbol).getBType();
         }
 
-        return symTable.noType;
-    }
-
-    private BType getCompletionBType(TypeSymbol completionType) {
-        if (completionType instanceof AbstractTypeSymbol) {
-            return ((AbstractTypeSymbol) completionType).getBType();
-        }
-
-        return symTable.nilType;
+        throw new IllegalArgumentException("Valid type parameter of Map type should be provided");
     }
 }
