+++ tests/ballerina-compiler-api-test/src/test/java/io/ballerina/semantic/api/test/TypeBuildersTest.java
@@ -20,7 +20,7 @@ package io.ballerina.semantic.api.test;
 
 import io.ballerina.compiler.api.SemanticModel;
 import io.ballerina.compiler.api.Types;
-import io.ballerina.compiler.api.impl.types.TypeBuilder;
+import io.ballerina.compiler.api.TypeBuilder;
 import io.ballerina.compiler.api.symbols.ArrayTypeSymbol;
 import io.ballerina.compiler.api.symbols.ErrorTypeSymbol;
 import io.ballerina.compiler.api.symbols.FunctionTypeSymbol;
